[MOD] Updated lesson 1 solution
